refactor: use base class
